Render dummy icon for Pokemon missing from icon set
The dummy icon is taken from the bottom right corner of the icon set
image.

The best way to take advantage of this is to fill all empty slots in the
last row
of the image with copies of the dummy icon (e.g. a question mark), or
if no empty slots, add a new row of dummy icons.