Overscan 1° beyond the ends of the level arc to correctly handle extreme marker locations
At level 1, and at (playerLevel+2), the marker hangs off the ends of the
level arc.
With the new averaging algorithm, this was causing failed detection of
Pokemon at (playerLevel+2). Such high levels aren't normally reachable
in the game, but they can/could be reached by exploiting glitches.

This issue doesn't show up for level 1 Pokemon because the jump from
level 1 to 1.5 on the arc is so massive.